fix_mssql: Split headers
