+++ CMakeLists.txt
@@ -1937,6 +1937,7 @@ set(WINDOWS_PLUGIN_FILES
         src/collectors/windows.plugin/GetSystemRAM.c
         src/collectors/windows.plugin/GetSystemCPU.c
         src/collectors/windows.plugin/NetdataMSSQLAPI.c
+        src/collectors/windows.plugin/NetdataMSSQLAPI.h
         src/collectors/windows.plugin/perflib-adcs.c
         src/collectors/windows.plugin/perflib-adfs.c
         src/collectors/windows.plugin/perflib-rrd.c
