+++ src/collectors/windows.plugin/windows_plugin.h
@@ -11,22 +11,9 @@
 // 2^24
 #define WINDOWS_MAX_KERNEL_OBJECT 16777216
 
-#include <sql.h>
-#include <sqlext.h>
+#include "NetdataMSSQLAPI.h"
 
 void *win_plugin_main(void *ptr);
-
-struct netdata_mssql_conn {
-    const char *hostname;
-    const char *username;
-    const char *password;
-    int port;
-};
-
-SQLHENV netdata_MSSQL_initialize_env();
-SQLHDBC netdata_MSSQL_start_connection(SQLHENV hEnv, char *dbconnstr);
-void netdata_MSSQL_cleanup_env(SQLHENV hEnv);
-void netdata_MSSQL_close_connection(SQLHDBC netdataDBC);
 
 extern char windows_shared_buffer[8192];
